+++ balance/weighting_methods/rake.py
@@ -6,6 +6,11 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import logging
+
+import math
+from fractions import Fraction
+
+from functools import reduce
 from typing import Callable, Dict, List, Union
 
 import numpy as np
@@ -249,3 +254,233 @@ def rake(
             # TODO: fix functions that use the perf and remove it from here
         },
     }
+
+
+def _lcm(a: int, b: int) -> int:
+    """
+    Calculates the least common multiple (LCM) of two integers.
+
+    The least common multiple (LCM) of two or more numbers is the smallest positive integer that is divisible by each of the given numbers.
+    In other words, it is the smallest multiple that the numbers have in common.
+    The LCM is useful when you need to find a common denominator for fractions or synchronize repeating events with different intervals.
+
+    For example, let's find the LCM of 4 and 6:
+
+    The multiples of 4 are: 4, 8, 12, 16, 20, ...
+    The multiples of 6 are: 6, 12, 18, 24, 30, ...
+    The smallest multiple that both numbers have in common is 12, so the LCM of 4 and 6 is 12.
+
+    The calculation is based on the property that the product of two numbers is equal to the product of their LCM and GCD: a * b = LCM(a, b) * GCD(a, b).
+    (proof: https://math.stackexchange.com/a/589299/1406)
+
+    Args:
+        a: The first integer.
+        b: The second integer.
+
+    Returns:
+        The least common multiple of the two integers.
+    """
+    # NOTE: this function uses math.gcd which calculates the greatest common divisor (GCD) of two integers.
+    # The greatest common divisor (GCD) of two or more integers is the largest positive integer that divides each of the given integers without leaving a remainder.
+    # In other words, it is the largest common factor of the given integers.
+    # For example, the GCD of 12 and 18 is 6, since 6 is the largest integer that divides both 12 and 18 without leaving a remainder.
+    # Similarly, the GCD of 24, 36, and 48 is 12, since 12 is the largest integer that divides all three of these numbers without leaving a remainder.
+    return abs(a * b) // math.gcd(a, b)
+
+
+def _proportional_array_from_dict(
+    input_dict: Dict[str, float], max_length: int = 10000
+) -> List[str]:
+    """
+    Generates a proportional array based on the input dictionary.
+
+    Args:
+        input_dict (Dict[str, float]): A dictionary where keys are strings and values are their proportions (float).
+        max_length (int): check if the length of the output exceeds the max_length. If it does, it will be scaled down to that length. Default is 10k.
+
+    Returns:
+        A list of strings where each key is repeated according to its proportion.
+
+    Examples:
+        ::
+            _proportional_array_from_dict({"a":0.2, "b":0.8})
+                # ['a', 'b', 'b', 'b', 'b']
+            _proportional_array_from_dict({"a":0.5, "b":0.5})
+                # ['a', 'b']
+            _proportional_array_from_dict({"a":1/3, "b":1/3, "c": 1/3})
+                # ['a', 'b', 'c']
+            _proportional_array_from_dict({"a": 3/8, "b": 5/8})
+                # ['a', 'a', 'a', 'b', 'b', 'b', 'b', 'b']
+            _proportional_array_from_dict({"a":3/5, "b":1/5, "c": 2/10})
+                # ['a', 'a', 'a', 'b', 'c']
+    """
+    # Filter out items with zero value
+    filtered_dict = {k: v for k, v in input_dict.items() if v > 0}
+
+    # Normalize the values if they don't sum to 1
+    sum_values = sum(filtered_dict.values())
+    if sum_values != 1:
+        filtered_dict = {k: v / sum_values for k, v in filtered_dict.items()}
+
+    # Convert the proportions to fractions
+    fraction_dict = {
+        k: Fraction(v).limit_denominator() for k, v in filtered_dict.items()
+    }
+
+    # Find the least common denominator (LCD) of the fractions
+    lcd = 1
+    for fraction in fraction_dict.values():
+        lcd *= fraction.denominator // math.gcd(lcd, fraction.denominator)
+
+    # Calculate the scaling factor based on max_length
+    scaling_factor = min(1, max_length / lcd)
+
+    result = []
+    for key, fraction in fraction_dict.items():
+        # Calculate the count for each key based on its proportion and scaling_factor
+        k_count = round((fraction * lcd).numerator * scaling_factor)
+        # Extend the result array with the key repeated according to its count
+        result.extend([key] * k_count)
+
+    return result
+
+
+def _find_lcm_of_array_lengths(arrays: Dict[str, List[str]]) -> int:
+    """
+    Finds the least common multiple (LCM) of the lengths of arrays in the input dictionary.
+
+    Args:
+        arrays: A dictionary where keys are strings and values are lists of strings.
+
+    Returns:
+        The LCM of the lengths of the arrays in the input dictionary.
+
+    Example:
+        ::
+            arrays = {
+                        "v1": ["a", "b", "b", "c"],
+                        "v2": ["aa", "bb"]
+                    }
+            _find_lcm_of_array_lengths(arrays)
+                # 4
+
+            arrays = {
+                        "v1": ["a", "b", "b", "c"],
+                        "v2": ["aa", "bb"],
+                        "v3": ["a1", "a2", "a3"]
+                    }
+            _find_lcm_of_array_lengths(arrays)
+                # 12
+    """
+    array_lengths = [len(arr) for arr in arrays.values()]
+    lcm_length = reduce(lambda x, y: _lcm(x, y), array_lengths)
+    return lcm_length
+
+
+def _realize_dicts_of_proportions(
+    dict_of_dicts: Dict[str, Dict[str, float]]
+) -> Dict[str, List[str]]:
+    """
+    Generates proportional arrays of equal length for each input dictionary.
+
+    This can be used to get an input dict of proportions of values, and produce a dict with arrays that realizes these proportions.
+    It can be used as input to the Sample object so it could be used for running raking.
+
+    Args:
+        dict_of_dicts: A dictionary of dictionaries, where each key is a string and
+                   each value is a dictionary with keys as strings and values as their proportions (float).
+
+    Returns:
+        A dictionary with the same keys as the input and equal length arrays as values.
+
+    Examples:
+        ::
+            dict_of_dicts = {
+                "v1": {"a": 0.2, "b": 0.6, "c": 0.2},
+                "v2": {"aa": 0.5, "bb": 0.5}
+            }
+
+            realize_dicts_of_proportions(dict_of_dicts)
+            # {'v1': ['a', 'b', 'b', 'b', 'c', 'a', 'b', 'b', 'b', 'c'], 'v2': ['aa', 'bb', 'aa', 'bb', 'aa', 'bb', 'aa', 'bb', 'aa', 'bb']}
+
+
+            dict_of_dicts = {
+                "v1": {"a": 0.2, "b": 0.6, "c": 0.2},
+                "v2": {"aa": 0.5, "bb": 0.5},
+                "v3": {"A": 0.2, "B": 0.8},
+            }
+            realize_dicts_of_proportions(dict_of_dicts)
+                # {'v1': ['a', 'b', 'b', 'b', 'c', 'a', 'b', 'b', 'b', 'c'],
+                #  'v2': ['aa', 'bb', 'aa', 'bb', 'aa', 'bb', 'aa', 'bb', 'aa', 'bb'],
+                #  'v3': ['A', 'B', 'B', 'B', 'B', 'A', 'B', 'B', 'B', 'B']}
+            # The above example could have been made shorter. But that's a limitation of the function.
+
+            dict_of_dicts = {
+                "v1": {"a": 0.2, "b": 0.6, "c": 0.2},
+                "v2": {"aa": 0.5, "bb": 0.5},
+                "v3": {"A": 0.2, "B": 0.8},
+                "v4": {"A": 0.1, "B": 0.9},
+            }
+            realize_dicts_of_proportions(dict_of_dicts)
+                # {'v1': ['a', 'b', 'b', 'b', 'c', 'a', 'b', 'b', 'b', 'c'],
+                #  'v2': ['aa', 'bb', 'aa', 'bb', 'aa', 'bb', 'aa', 'bb', 'aa', 'bb'],
+                #  'v3': ['A', 'B', 'B', 'B', 'B', 'A', 'B', 'B', 'B', 'B'],
+                #  'v4': ['A', 'B', 'B', 'B', 'B', 'B', 'B', 'B', 'B', 'B']}
+    """
+    # Generate proportional arrays for each dictionary
+    arrays = {k: _proportional_array_from_dict(v) for k, v in dict_of_dicts.items()}
+
+    # Find the LCM over the lengths of all the arrays
+    lcm_length = _find_lcm_of_array_lengths(arrays)
+
+    # Extend each array to have the same LCM length while maintaining proportions
+    result = {}
+    for k, arr in arrays.items():
+        factor = lcm_length // len(arr)
+        extended_arr = arr * factor
+        result[k] = extended_arr
+
+    return result
+
+
+def prepare_marginal_dist_for_raking(
+    dict_of_dicts: Dict[str, Dict[str, float]]
+) -> pd.DataFrame:
+    """
+    Realizes a nested dictionary of proportions into a DataFrame.
+
+    Args:
+        dict_of_dicts: A nested dictionary where the outer keys are column names
+                   and the inner dictionaries have keys as category labels
+                   and values as their proportions (float).
+
+    Returns:
+        A DataFrame with columns specified by the outer keys of the input dictionary
+        and rows containing the category labels according to their proportions.
+        An additional "id" column is added with integer values as row identifiers.
+
+    Example:
+        ::
+            print(prepare_marginal_dist_for_raking({
+                            "A": {"a": 0.5, "b": 0.5},
+                            "B": {"x": 0.2, "y": 0.8}
+                        }))
+            # Returns a DataFrame with columns A, B, and id
+                #    A  B  id
+                # 0  a  x   0
+                # 1  b  y   1
+                # 2  a  y   2
+                # 3  b  y   3
+                # 4  a  y   4
+                # 5  b  x   5
+                # 6  a  y   6
+                # 7  b  y   7
+                # 8  a  y   8
+                # 9  b  y   9
+    """
+    target_dict_from_marginals = _realize_dicts_of_proportions(dict_of_dicts)
+    target_df_from_marginals = pd.DataFrame.from_dict(target_dict_from_marginals)
+    # Add an id column:
+    target_df_from_marginals["id"] = range(target_df_from_marginals.shape[0])
+
+    return target_df_from_marginals