+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ==================
 ### New Features
 - Add `rake` method to .adjust (currently in beta, given that it doesn't handles marginal target as input).
+- Add a new function `prepare_marginal_dist_for_raking` - to take in a dict of marginal proportions and turn them into a pandas DataFrame. This can serve as an input target population for raking.
+
 
 0.7.0 (2023-04-10)
 ==================